refactor: move Slice.At tests into slice_test.go
Per user feedback, the tests for Slice.At should not be in a separate
file. This commit consolidates them into the main slice_test.go file.
